+++ app/src/main/java/com/adammcneilly/reader/ui/theme/Color.kt
@@ -1,6 +1,6 @@
 @file:Suppress("MagicNumber")
 
-package template.theme
+package com.adammcneilly.reader.ui.theme
 
 import androidx.compose.ui.graphics.Color
