+++ buildscripts/setup.gradle
@@ -1,8 +1,8 @@
 def renameConfig = [
         templateName  : "template",
         templateAppId : "template.app.id",
-        newPackage    : "domain.yourname.app",
-        newProjectName: "Your Project",
+        newPackage    : "com.adammcneilly.reader",
+        newProjectName: "Reader",
 ]
 
 task renameAppPackage(type: Copy) {
